+++ src/rules/declaration-colon-newline-after/__tests__/index.js
@@ -28,6 +28,10 @@ testRule("always", tr => {
     "a { color\r\n:\r\npink }",
     "CRLF before and after"
   )
+  tr.ok(
+    "$map: (key: value)",
+    "SCSS map with no newlines"
+  )
 
   tr.notOk(
     "a { color :pink; }",
@@ -82,6 +86,11 @@ testRule("always-multi-line", tr => {
     "    0 0 0 1px #5b9dd9\n" +
     "    0 0 2px 1px rgba(30, 140, 190, 0.8);\n" +
     "}"
+  )
+
+  tr.ok(
+    "$map\n: (\nkey: value,\nkey2 :value2)",
+    "SCSS map with newlines"
   )
 
   tr.notOk(